GA optimization time estimation + other small fixes.

+++ trunk/web-client/src/main/java/eu/cityopt/controller/ControllerService.java
@@ -584,7 +584,7 @@ public class ControllerService {
 			} else {
 			    Duration d = timeEstimatorService.predictSimulationRuntime(projectId, scenarioId);
 			    if (d != null) {
-                    model.put("simulationEstimate", (d.getSeconds() / 60 + 1) + " min");
+                    model.put("simulationEstimate", (int)(d.getSeconds() / 60 + 1) + " min");
 			    }
 			}
